Create Sales Tax Calculator 2 10-21-22
Way more detailed then the last sales tax calculator, may be a bit too wordy and long but def will work on good definitions in the future. Messed around a bit with lambda functions in the building of this project, but realized it was far easier to write without lambda.